requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
letta-client==1.3.1
firebase-admin==7.1.0
requests==2.32.5
gunicorn==23.0.0